Drop special-casing of "fragile" macros to be preserved
Simply because one wants to preserve from expanding anything (actually
non-parameterized macros only -- for that another change is upcoming)
not explicitly blacklisted using distill-spec-prefix and except for
macros/builtins enumerated in the re_noexpand precooked regex (that is
actually covering a great deal of alike tokens in this context already),
not just the picked ones.

As a bonus, this simplifies the code.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>